requirements.txt
google-genai>=1.0.0
openai>=1.50.0
higgsfield-client>=0.1.0
python-dotenv>=1.0.1
rich>=13.7.0
httpx>=0.27.0
